requirements.txt
# AI Handwriting Recognition System - Dependencies
# Core ML Libraries
tensorflow>=2.12.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# GUI Framework
customtkinter>=5.2.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: For advanced features
# torch>=2.0.0  # PyTorch alternative
# torchvision>=0.15.0
